Added functionality for 2 card flip
Added function that allows the user to flip 2 cards, and when match is false, then it flips them back over by removing the class.

The struggle I had with this was that I reversed my push parameters. My array was empty. Now, I will use the forEach method instead of a loop